requirements.txt
python-dotenv
requests
agno 